Refactor: Integer -> Long 타입 변경 대응
- 금액 정보를 Integer -> Long으로 타입 변경

+++ semicolon/src/main/java/dukku/semicolon/shared/deposit/dto/DepositHistoryResponse.java
@@ -38,8 +38,8 @@ public class DepositHistoryResponse {
     public static class DepositHistoryHistoryItem {
         private String depositHistoryId; // UUID 형태의 String (이력 ID)
         private String type; // 변동 유형 (Enum String)
-        private Integer amount; // 변동 금액
-        private Integer balanceAfter; // 변동 후 잔액
+        private Long amount; // 변동 금액
+        private Long balanceAfter; // 변동 후 잔액
         private ReferenceInfo ref; // 참조 정보 (결제/주문 등)
         private OffsetDateTime createdAt; // 생성 일시
     }
